requirements.txt
-i https://pypi.org/simple
click==7.1.1
flask==1.1.1
itsdangerous==1.1.0
jinja2==2.11.1
markupsafe==1.1.1
numpy==1.18.2
pillow==7.0.0
six==1.14.0
torch==1.4.0
torchvision==0.5.0
werkzeug==1.0.0
